+++ src/DB/Migrations/MessageMigration01.php
@@ -12,21 +12,22 @@
 	class MessageMigration01 {
 		public static function Up( ) {
 			$table_name = 'messages';
-			Capsule::schema()->create($table_name, function ($table) {
-				$table->bigIncrements('id');
-				$table->unsignedBigInteger('gateway_id');
-				$table->string('senderID');
-				$table->string('recipientsCount');
-				$table->string('status');
-				$table->string('response');
-				$table->string('groups');
-				$table->text('message');
-				$table->string('time')->nullable();
-				$table->json('metaData')->nullable();
-				$table->foreign('gateway_id')->references('id')->on('gateways')->onDelete('CASCADE');
-				$table->timestamps();
-			});
-			
+			if ( !Capsule::schema()->hasTable( $table_name ) ) {
+				Capsule::schema()->create( $table_name, function ( $table ) {
+					$table->bigIncrements( 'id' );
+					$table->unsignedBigInteger( 'gateway_id' );
+					$table->string( 'senderID' );
+					$table->string( 'recipientsCount' );
+					$table->string( 'status' );
+					$table->string( 'response' );
+					$table->string( 'groups' );
+					$table->text( 'message' );
+					$table->string( 'time' )->nullable();
+					$table->json( 'metaData' )->nullable();
+					$table->foreign( 'gateway_id' )->references( 'id' )->on( 'gateways' )->onDelete( 'CASCADE' );
+					$table->timestamps();
+				} );
+			}
 		}
 		
 		public static function Down(  ) {
